Store npmts as int not double

+++ DataModel/DataModel.h
@@ -78,8 +78,8 @@ class DataModel {
 
   double IDPMTDarkRate;
   double ODPMTDarkRate;
-  double IDNPMTs;
-  double ODNPMTs;
+  int IDNPMTs;
+  int ODNPMTs;
   
   WCSimRootOptions WCSimOpt;
   WCSimRootEvent   WCSimEvt;
